@@ -1,12 +1,35 @@
 [build-system]
-requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+requires = ["poetry-core>=2.0"]
 
-[tool.poetry]
+[dependency-groups]
+dev = [
+  "black>=25.1.0",
+  "pylint>=3.3.0",
+  "pytest>=8.3.0",
+  "ruff>=0.9.0",
+  "toml-sort>=0.24.0",
+  "yamllint>=1.36.0",
+]
+docs = [
+  "black>=25.1.0",
+  "mkdocs>=1.6.0",
+  "mkdocs-literate-nav>=0.6.0",
+  "mkdocs-material>=9.6.0",
+  "mkdocstrings[python]>=0.29.0",
+]
+lab = [
+  "ipykernel>=6.29.0",
+  "ipympl>=0.9.0",
+  "jupyterlab>=4.3.0",
+  "matplotlib>=3.10.0",
+]
+
+[project]
 authors = [
-  "Cristina Rea <crea@psfc.mit.edu>",
-  "Gregorio L. Trevisan <gtrevisan@psfc.mit.edu>",
-  "MIT PSFC Disruption Studies Group <disruption-py@lists.psfc.mit.edu>",
+  {name = "Cristina Rea", email = "crea@psfc.mit.edu"},
+  {name = "Gregorio L. Trevisan", email = "gtrevisan@psfc.mit.edu"},
+  {name = "MIT PSFC Disruption Studies Group", email = "disruption-py@lists.psfc.mit.edu"},
 ]
 classifiers = [
   "Development Status :: 4 - Beta",
@@ -22,9 +45,18 @@ classifiers = [
   "Topic :: Software Development :: Libraries :: Python Modules",
   "Topic :: Software Development :: Version Control :: Git",
 ]
+dependencies = [
+  "dynaconf>=3.2.0",
+  "loguru>=0.7.0",
+  "numpy>=1.26.0,<2.0.0",
+  "pandas[hdf5]>=2.2.0",
+  "pyodbc>=5.2.0",
+  "scikit_learn>=1.6.0",
+  "scipy>=1.15.0",
+  "SQLAlchemy>=2.0.0",
+  "tqdm>=4.67.0",
+]
 description = "An open-source physics-based Scientific Framework for Disruption Analysis of Fusion Plasmas for AI/ML applications"
-documentation = "https://mit-psfc.github.io/disruption-py/"
-homepage = "https://disruptions.mit.edu/"
 keywords = [
   "plasma physics",
   "nuclear fusion",
@@ -33,45 +65,41 @@ keywords = [
 ]
 license = "MIT"
 maintainers = [
-  "Gregorio L. Trevisan <gtrevisan@psfc.mit.edu>",
+  {name = "Gregorio L. Trevisan", email = "gtrevisan@psfc.mit.edu"},
 ]
 name = "disruption-py"
 readme = "README.md"
-repository = "https://github.com/MIT-PSFC/disruption-py/"
-version = "0.9.0"
+requires-python = ">=3.11,<4"
+version = "0.10.0"
 
-[tool.poetry.dependencies]
-python = "^3.10"
-dynaconf = "^3.2.0"
-loguru = "^0.7.0"
-numpy = "^1.26.0"
-pandas = {extras = ["hdf5"], version = "^2.2.0"}
-pyodbc = "^5.2.0"
-scikit_learn = "^1.6.0"
-scipy = "^1.15.0"
-SQLAlchemy = "^2.0.0"
-tqdm = "^4.67.0"
+[project.scripts]
+disruption-workflow = "examples.simple:cli"
+
+[project.urls]
+documentation = "https://mit-psfc.github.io/disruption-py/"
+homepage = "https://disruptions.mit.edu/"
+repository = "https://github.com/MIT-PSFC/disruption-py/"
 
 [tool.poetry.group.dev]
 optional = true
 
 [tool.poetry.group.dev.dependencies]
-black = "^24.10.0"
+black = "^25.1.0"
 pylint = "^3.3.0"
 pytest = "^8.3.0"
 ruff = "^0.9.0"
 toml-sort = "^0.24.0"
-yamllint = "^1.35.0"
+yamllint = "^1.36.0"
 
 [tool.poetry.group.docs]
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-black = "^24.10.0"
+black = "^25.1.0"
 mkdocs = "^1.6.0"
 mkdocs-literate-nav = "^0.6.0"
-mkdocs-material = "^9.5.0"
-mkdocstrings = {extras = ["python"], version = "^0.27.0"}
+mkdocs-material = "^9.6.0"
+mkdocstrings = {extras = ["python"], version = "^0.29.0"}
 
 [tool.poetry.group.lab]
 optional = true
@@ -82,12 +110,7 @@ ipympl = "^0.9.0"
 jupyterlab = "^4.3.0"
 matplotlib = "^3.10.0"
 
-[tool.poetry.scripts]
-disruption-workflow = "examples.simple:cli"
-
 [tool.pylint]
-extension-pkg-allow-list = ["pyodbc"]
-ignored-modules = ["MDSplus"]
 disable = [
   "fixme",
   "too-few-public-methods",
@@ -99,3 +122,14 @@ disable = [
   "too-many-positional-arguments",
   "too-many-statements",
 ]
+extension-pkg-allow-list = ["pyodbc"]
+ignored-modules = ["MDSplus"]
+
+[tool.tomlsort]
+ignore_case = true
+in_place = true
+sort_table_keys = true
+trailing_comma_inline_array = true
+
+[tool.uv]
+default-groups = []
