@@ -4,6 +4,8 @@
 This module provides a decorator to signify methods that calculate physics quantities.
 """
 
+import time
+from functools import wraps
 from typing import Callable, List, Union
 
 from disruption_py.core.physics_method.caching import cache_method
@@ -14,7 +16,6 @@ from disruption_py.machine.tokamak import Tokamak
 def physics_method(
     cache: bool = True,
     tokamak: Union[Tokamak, List[Tokamak]] = None,
-    tags: List[str] = None,
     columns: Union[List[str], Callable] = None,
 ) -> Callable:
     """
@@ -23,7 +24,7 @@ def physics_method(
     The decorated method calculates disruption parameters and returns a Pandas
     DataFrame. All decorated methods must take the single argument params of type
     `PhysicsMethodParams`. The decorated method will be run if designated by the
-    `run_methods`, `run_tags`, or `run_columns` attributes of the `RetrievalSettings`
+    `run_methods` or `run_columns` attributes of the `RetrievalSettings`
     class, and if included inside of the `custom_physics_methods` argument of the
     `retrieval_settings` or in the built-in method list. If run, the result of the
     decorated method will be output to the `output_setting`.
@@ -38,10 +39,6 @@ def physics_method(
     tokamak : Union['Tokamak', List['Tokamak']], optional
         A list of Tokamak objects that represent which tokamak this method may
         be used for, by default None, allows the method to be run for any tokamak.
-    tags : List[str], optional
-        The list of tags to help identify this method. Tags can be used when calling
-        `get_shots_data` to have disruption_py use multiple functions at the same
-        time. Default is ["all"].
     columns : Union[List[str], Callable], optional
         The columns that are in the DataFrame returned by the method. Alternatively,
         you can pass a method that returns the names of used trees at runtime. Default
@@ -59,15 +56,27 @@ def physics_method(
         else:
             wrapper = method
 
+        # Add elapsed time log
+        @wraps(wrapper)
+        def timed_wrapper(params, *args, **kwargs):
+            start_time = time.time()
+            result = wrapper(params, *args, **kwargs)
+            params.logger.verbose(
+                "{t:.3f}s : {name}",
+                name=method.__name__,
+                t=time.time() - start_time,
+            )
+            return result
+
         method_metadata = MethodMetadata(
             name=method.__name__,
             cache=cache,
             tokamaks=tokamak,
             columns=columns,
-            tags=tags,
         )
 
-        wrapper.method_metadata = method_metadata
-        return wrapper
+        timed_wrapper.method_metadata = method_metadata
+
+        return timed_wrapper
 
     return outer_wrapper
