@@ -144,7 +144,7 @@ def filter_methods_to_run(
     all_bound_method_metadata : list[BoundMethodMetadata]
         A list of bound method metadata instances.
     retrieval_settings : RetrievalSettings
-        The settings that dictate which methods should be run based on tags, methods,
+        The settings that dictate which methods should be run based on methods
         and columns.
     physics_method_params : PhysicsMethodParams
         The parameters that will be passed to the methods that are run.
@@ -154,10 +154,12 @@ def filter_methods_to_run(
     list
         A list of bound method metadata instances that are eligible to run.
     """
-    tags = retrieval_settings.run_tags
     methods = retrieval_settings.run_methods
-    columns = REQUIRED_COLS.union(retrieval_settings.run_columns)
-
+    if retrieval_settings.run_columns is not None:
+        columns = REQUIRED_COLS.union(retrieval_settings.run_columns)
+    else:
+        columns = None
+    only_excluded_methods_specified = all("~" in method for method in methods or [])
     methods_to_run = []
     for bound_method_metadata in all_bound_method_metadata:
         # exclude if tokamak does not match
@@ -171,15 +173,27 @@ def filter_methods_to_run(
         ):
             continue
 
-        if tags is not None and bool(
-            set(bound_method_metadata.tags).intersection(tags)
-        ):
-            methods_to_run.append(bound_method_metadata)
-        elif methods is not None and bound_method_metadata.name in methods:
-            methods_to_run.append(bound_method_metadata)
-        elif columns is not None and bool(
+        both_none = methods is None and columns is None
+        method_specified = methods is not None and bound_method_metadata.name in methods
+        column_speficied = columns is not None and bool(
             set(bound_method_metadata.columns).intersection(columns)
-        ):
+        )
+        is_not_excluded = (
+            only_excluded_methods_specified
+            and not columns
+            and methods
+            and (("~" + bound_method_metadata.name) not in methods)
+        )
+        should_run = (
+            both_none or method_specified or column_speficied or is_not_excluded
+        )
+
+        # reasons that methods should be exluded from should run
+        should_not_run = (
+            methods is not None and ("~" + bound_method_metadata.name) in methods
+        )
+
+        if should_run and not should_not_run:
             methods_to_run.append(bound_method_metadata)
         else:
             physics_method_params.logger.debug(
@@ -207,7 +221,6 @@ def populate_method(
     Any
         The result of the executed method, or None if an error occurred.
     """
-    start_time = time.time()
     method = bound_method_metadata.bound_method
     name = bound_method_metadata.name
 
@@ -239,11 +252,6 @@ def populate_method(
         physics_method_params.logger.opt(exception=True).debug(e)
         result = {col: [np.nan] for col in bound_method_metadata.columns}
 
-    physics_method_params.logger.verbose(
-        "{t:.3f}s : {name}",
-        name=name,
-        t=time.time() - start_time,
-    )
     return result
 
 
@@ -256,7 +264,7 @@ def populate_shot(
 
     This function executes the physics methods included through the
     `custom_physics_methods` property of retrieval_settings or in the built-in list
-    of methods. It selects methods based on run_methods, run_tags, and run_columns
+    of methods. It selects methods based on run_methods and run_columns
     in retrieval_settings.
 
     Parameters
@@ -375,7 +383,10 @@ def populate_shot(
 
     local_data = pd.concat([pre_filled_shot_data] + filtered_methods, axis=1)
     local_data = local_data.loc[:, ~local_data.columns.duplicated()]
-    if retrieval_settings.only_requested_columns:
+    if (
+        retrieval_settings.only_requested_columns
+        and retrieval_settings.run_columns is not None
+    ):
         include_columns = list(
             REQUIRED_COLS.union(
                 set(retrieval_settings.run_columns).intersection(
