@@ -1094,7 +1094,7 @@ class D3DPhysicsMethods:
         rho_vn_diag_almost = interp1(
             efit_dict["time"], efit_dict["rhovn"], ts["time"], axis=0
         )
-        rho_vn_diag = np.empty(psin.shape[:2])
+        rho_vn_diag = np.full(psin.shape[:2], np.nan)
         # Ger the implied psin grid for rhovn
         psin_interp = np.linspace(0, 1, efit_dict["rhovn"].shape[1])
         # Interpolate again to get rhovn on same psin base
