@@ -23,13 +23,15 @@ def main():
     elif tokamak is Tokamak.CMOD:
         shotlist = [1150805012]
         shape = (62, 23)
+    elif tokamak is Tokamak.EAST:
+        shotlist = [55555]
+        shape = (69, 17)
     else:
         raise ValueError(f"Unspecified or unsupported tokamak: {tokamak}.")
 
     print(f"Initialized for tokamak: {tokamak.value}")
 
     retrieval_settings = RetrievalSettings(
-        run_tags=[],
         run_methods=run_methods,
         efit_nickname_setting="default",
     )
