@@ -14,7 +14,10 @@ import pytest
 from disruption_py.config import config
 from disruption_py.core.utils.misc import without_duplicates
 from disruption_py.inout.sql import ShotDatabase
-from disruption_py.settings.output_setting import SQLOutputSetting
+from disruption_py.settings.output_setting import (
+    BatchedCSVOutputSetting,
+    SQLOutputSetting,
+)
 from disruption_py.settings.retrieval_settings import RetrievalSettings
 from disruption_py.workflow import get_database, get_shots_data
 from tests.conftest import skip_on_fast_execution
@@ -42,8 +45,7 @@ def setup_shot_database(shotlist, shot_database):
     This fixture automatically removes data for each shot in the
     provided shotlist from the shot_database.
     """
-    for shot in shotlist:
-        shot_database.remove_shot_data(shot)
+    shot_database.remove_shot_data(shotlist)
 
 
 @pytest.fixture(scope="module", name="initial_mdsplus_data")
@@ -61,7 +63,6 @@ def initial_mdsplus_data_fixture(shotlist, tokamak, test_file_path_f) -> Dict:
     retrieval_settings = RetrievalSettings(
         efit_nickname_setting="disruption",
         run_columns=FIRST_ITERATION_COLUMNS,
-        run_tags=[],
         only_requested_columns=True,
     )
     all_outputs = get_shots_data(
@@ -103,25 +104,36 @@ def test_output_exists(initial_mdsplus_data, test_file_path_f):
 
 @skip_on_fast_execution
 def test_sql_output_setting(
-    shotlist, shot_database: ShotDatabase, initial_mdsplus_data_df
+    shotlist, shot_database: ShotDatabase, initial_mdsplus_data_df, test_file_path_f
 ) -> Dict:
     """
     Ensure SQL output setting works by reading from an initial writeback to
     SQL, then by updating the SQL with another writeback, and making sure data from
     MDSplus matches the data from SQL for both retrievals.
     """
-    # Test initial database readback
-    result = shot_database.get_shots_data(shotlist, sql_table=WRITE_DATABASE_TABLE_NAME)
-    assert_frame_equal_unordered(
-        result[FIRST_ITERATION_COLUMNS],
-        initial_mdsplus_data_df[FIRST_ITERATION_COLUMNS],
+
+    # Test initial database writeback
+    result = shot_database.get_shots_data(
+        shotlist=shotlist,
+        cols=FIRST_ITERATION_COLUMNS,
+        sql_table=WRITE_DATABASE_TABLE_NAME,
     )
+    try:
+        assert_frame_equal_unordered(
+            result[FIRST_ITERATION_COLUMNS],
+            initial_mdsplus_data_df[FIRST_ITERATION_COLUMNS],
+        )
+    except AssertionError as e:
+        result.to_csv(test_file_path_f("-1L.csv"))
+        initial_mdsplus_data_df[FIRST_ITERATION_COLUMNS].to_csv(
+            test_file_path_f("-1R.csv")
+        )
+        raise AssertionError(f"First writeback to SQL failed! {e.args[0]}") from e
 
     # Do second retrieval that updates the data for the columns
     retrieval_settings = RetrievalSettings(
         efit_nickname_setting="disruption",
         run_columns=ALL_ITERATION_COLUMNS,
-        run_tags=[],
         only_requested_columns=True,
     )
     shot_data = get_shots_data(
@@ -129,11 +141,39 @@ def test_sql_output_setting(
         retrieval_settings=retrieval_settings,
         output_setting=SQLOutputSetting(
             table_name=WRITE_DATABASE_TABLE_NAME,
+            should_update=True,
             should_override_columns=SECOND_ITERATION_COLUMNS,
         ),
         num_processes=2,
     )
-    result = shot_database.get_shots_data(shotlist, sql_table=WRITE_DATABASE_TABLE_NAME)
-    assert_frame_equal_unordered(
-        result[ALL_ITERATION_COLUMNS], shot_data[ALL_ITERATION_COLUMNS]
+    result = shot_database.get_shots_data(
+        shotlist=shotlist,
+        cols=ALL_ITERATION_COLUMNS,
+        sql_table=WRITE_DATABASE_TABLE_NAME,
     )
+    try:
+        assert_frame_equal_unordered(
+            result[ALL_ITERATION_COLUMNS], shot_data[ALL_ITERATION_COLUMNS]
+        )
+    except AssertionError as e:
+        result[ALL_ITERATION_COLUMNS].to_csv(test_file_path_f("-2L.csv"))
+        shot_data[ALL_ITERATION_COLUMNS].to_csv(test_file_path_f("-2R.csv"))
+        raise AssertionError(f"Second writeback to SQL failed! {e.args[0]}") from e
+
+
+def test_batch_csv(tokamak, test_file_path_f, shotlist):
+    """
+    Test the batch csv output setting to ensure it outputs the same columns in
+    the same order as the dataframe in memory.
+    """
+    csv = test_file_path_f("-batch.csv")
+    out = get_shots_data(
+        tokamak=tokamak,
+        shotlist_setting=shotlist,
+        num_processes=2,
+        # Use a batch size less than the number of shots to ensure multiple batches
+        # are written to the CSV file.
+        output_setting=BatchedCSVOutputSetting(filepath=csv, batch_size=1),
+    )
+    df = pd.read_csv(csv)
+    assert_frame_equal_unordered(out, df)
