@@ -8,6 +8,7 @@ sources and the time domain of the data retrieved.
 import pandas as pd
 import pytest
 
+from disruption_py.core.physics_method.runner import REQUIRED_COLS
 from disruption_py.core.utils.misc import safe_df_concat
 from disruption_py.inout.mds import ProcessMDSConnection
 from disruption_py.machine.tokamak import Tokamak
@@ -52,7 +53,6 @@ def test_cache_setting_sql(tokamak, shotlist, num_processes):
         cache_setting="sql",
         use_cache_setting_timebase=True,
         run_columns=["time_until_disrupt"],
-        run_tags=[],
         only_requested_columns=True,
         efit_nickname_setting="default",
     )
@@ -128,8 +128,6 @@ def test_only_requested_columns(tokamak, shotlist):
     """
     retrieval_settings = RetrievalSettings(
         run_columns=["ip", "q95"],
-        run_tags=[],
-        run_methods=[],
         only_requested_columns=True,
     )
     results = get_shots_data(
@@ -151,8 +149,11 @@ def test_domain_setting(tokamak, shotlist, domain_setting, full_time_domain_data
     Test the two partial domain settings by comparing their start and end times
     with the full domain.
     """
-    if tokamak == Tokamak.D3D and domain_setting == "rampup_and_flattop":
-        pytest.skip("rampup_and_flattop domain setting not defined for DIII-D")
+    if (
+        tokamak in [Tokamak.D3D, Tokamak.EAST]
+        and domain_setting == "rampup_and_flattop"
+    ):
+        pytest.skip(f"{domain_setting} domain setting not defined for {tokamak.value}")
     retrieval_settings = RetrievalSettings(
         efit_nickname_setting="default", domain_setting=domain_setting
     )
@@ -175,3 +176,71 @@ def test_domain_setting(tokamak, shotlist, domain_setting, full_time_domain_data
             assert f_start < p_start < p_end <= f_end
         else:
             assert f_start == p_start < p_end < f_end
+
+
+@skip_on_fast_execution
+@pytest.mark.parametrize(
+    "run_methods, run_columns, expected_cols, forbidden_cols",
+    [
+        # Test run_methods with run_columns=None
+        (None, None, None, []),
+        ([], None, REQUIRED_COLS, []),
+        (["~get_kappa_area"], None, None, ["kappa_area"]),
+        (["get_kappa_area"], None, REQUIRED_COLS | {"kappa_area"}, []),
+        # Test run_columns with run_methods=None
+        (None, [], REQUIRED_COLS, []),
+        (None, ["kappa_area"], REQUIRED_COLS | {"kappa_area"}, []),
+        # Test run_methods and run_columns combo
+        ([], [], REQUIRED_COLS, []),
+        (["get_kappa_area"], [], REQUIRED_COLS | {"kappa_area"}, []),
+        (
+            ["get_kappa_area"],
+            ["greenwald_fraction"],
+            REQUIRED_COLS | {"kappa_area", "n_e", "dn_dt", "greenwald_fraction"},
+            [],
+        ),
+        (
+            ["~get_kappa_area"],
+            ["greenwald_fraction"],
+            REQUIRED_COLS | {"n_e", "dn_dt", "greenwald_fraction"},
+            [],
+        ),
+        (["~get_kappa_area"], ["kappa_area"], REQUIRED_COLS, []),
+    ],
+)
+def test_run_methods_and_columns(
+    tokamak,
+    shotlist,
+    run_methods,
+    run_columns,
+    expected_cols,
+    forbidden_cols,
+    full_time_domain_data,
+):
+    """
+    Test the `run_methods` and `run_columns` parameters of RetrievalSettings.
+
+    - If both are None, all methods are run
+    - If one is None or [], the methods/columns specified by the other are run
+    - If both are specified, the combined methods/columns are run
+    - If `run_methods` excludes a method returning a column specified in `run_columns`,
+      the method is not run
+    """
+    num_all_cols = len(full_time_domain_data[0].columns)
+    retrieval_settings = RetrievalSettings(
+        run_methods=run_methods,
+        run_columns=run_columns,
+    )
+    results = get_shots_data(
+        tokamak=tokamak,
+        shotlist_setting=shotlist,
+        retrieval_settings=retrieval_settings,
+        num_processes=2,
+        log_settings="WARNING",
+    )
+    # Expected columns None means all columns (except forbidden cols) are returned
+    if expected_cols is None:
+        assert len(results.columns) == num_all_cols - len(forbidden_cols)
+    else:
+        assert set(results.columns) == set(expected_cols)
+    assert not any(col in results.columns for col in forbidden_cols)
