@@ -18,6 +18,7 @@ from disruption_py.machine.tokamak import Tokamak
 from disruption_py.settings import LogSettings, RetrievalSettings
 from disruption_py.workflow import get_shots_data
 from tests.utils.data_difference import DataDifference
+from tests.utils.factory import get_tokamak_test_columns
 
 
 def get_fresh_data(
@@ -25,11 +26,25 @@ def get_fresh_data(
     shotlist: List[int],
     log_file_path: str,
     test_columns: List[str] = None,
+    console_log_level: str = "WARNING",
 ) -> Dict[int, pd.DataFrame]:
     """
     Get fresh data for a list of shots.
 
-     Returns
+    Parameters
+    ----------
+    tokamak : Tokamak
+        The tokamak for which to retrieve data.
+    shotlist : List[int]
+        A list of shot IDs to retrieve data for.
+    log_file_path : str
+        The path to the log file.
+    test_columns : List[str], optional
+        A list of columns to retrieve.
+    console_log_level : str, optional
+        The log level for console output. Default is "WARNING".
+
+    Returns
     -------
     Dict[int, pd.DataFrame]
         Dictionary mapping shot IDs to retrieved fresh data.
@@ -37,9 +52,8 @@ def get_fresh_data(
     retrieval_settings = RetrievalSettings(
         efit_nickname_setting="disruption",
         time_setting="disruption_warning",
-        run_tags=[] if test_columns else ["all"],
-        run_columns=test_columns if test_columns else [],
-        only_requested_columns=test_columns,
+        run_columns=test_columns,
+        only_requested_columns=True,
     )
     shot_data = get_shots_data(
         tokamak=tokamak,
@@ -51,7 +65,7 @@ def get_fresh_data(
             log_file_path=log_file_path,
             log_file_write_mode="w",
             file_log_level="DEBUG",
-            console_log_level="WARNING",
+            console_log_level=console_log_level,
         ),
     )
     return shot_data
@@ -164,7 +178,14 @@ def eval_shot_against_cache(
             failure=failure,
             mismatch_string=data_difference.column_mismatch_string,
         )
-    assert not data_difference.failed, "Comparison failed"
+    # Python tests should not assert for expected failures to only catch unexpected failures
+    # Pytest should assert for expected failures to confirm the test fails or
+    # to catch unexpected successes
+    if "PYTEST_CURRENT_TEST" in os.environ or not expect_failure:
+        assert not data_difference.failed, (
+            f"Comparison failed on shot {data_difference.shot_id}, "
+            "column {data_difference.data_column}"
+        )
 
     return data_difference
 
@@ -174,6 +195,7 @@ def eval_against_cache(
     shotlist: List[int],
     expected_failure_columns: List[str],
     test_columns=None,
+    console_log_level="WARNING",
 ) -> Dict[int, pd.DataFrame]:
     """
     Evaluate fresh data against cached data for specified shots.
@@ -193,6 +215,8 @@ def eval_against_cache(
     test_columns : List[str], optional
         A list of columns to test against the cached data. If None, the function
         will determine the columns based on the available data.
+    console_log_level : str, optional
+        The log level for console output. Default is "WARNING".
 
     Returns
     -------
@@ -219,19 +243,12 @@ def eval_against_cache(
             shotlist=shotlist,
             log_file_path=os.path.join(tempfolder, "data_retrieval.log"),
             test_columns=test_columns,
+            console_log_level=console_log_level,
         )
     cache_data = get_cached_from_fresh(tokamak, shotlist, fresh_data, test_columns)
 
     if test_columns is None:
-        fresh_columns = set().union(*(df.columns for df in fresh_data.values()))
-        cache_columns = set().union(*(df.columns for df in cache_data.values()))
-        # Handle when one of fresh/cache has no data
-        if len(fresh_columns) == 0:
-            test_columns = sorted(cache_columns)
-        elif len(cache_columns) == 0:
-            test_columns = sorted(fresh_columns)
-        else:
-            test_columns = sorted(fresh_columns.intersection(cache_columns))
+        test_columns = get_tokamak_test_columns(tokamak)
 
     data_differences = eval_shots_against_cache(
         shotlist=shotlist,
